pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "hedge"
version = "0.1.0"
description = "Automated momentum trading system for equities"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "yfinance>=0.2.30",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "lxml>=4.9",
    "html5lib>=1.1",
    "APScheduler>=3.10",
]

[project.optional-dependencies]
alpaca = ["alpaca-py>=0.20"]
ibkr = ["ib_insync>=0.9"]
ml = ["scikit-learn>=1.3", "lightgbm>=4.0", "xgboost>=2.0"]
dashboard = ["streamlit>=1.28", "plotly>=5.18"]
dev = ["pytest>=7.4", "pytest-cov>=4.1", "ruff>=0.1"]
all = ["hedge[alpaca,ml,dashboard,dev]"]

[project.scripts]
hedge = "hedge.cli:main"

[tool.setuptools.packages.find]
include = ["hedge*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
